% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert}
\alias{convert}
\title{Convert Between File Formats}
\usage{
convert(
  file,
  from = c("csv2", "csv", "delim2", "delim", "dta"),
  to = c("csv", "csv2", "delim2", "delim", "dta"),
  ...
)
}
\arguments{
\item{file}{File to convert.}

\item{from}{Input file format.}

\item{to}{Output file format.}

\item{...}{Additional arguments passed to reading/writing functions.}
}
\value{
None.
}
\description{
Reads a file in one format and saves it in another.
}
