% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropbox}
\alias{dropbox}
\title{Path to Dropbox Folder}
\usage{
dropbox(dir)
}
\arguments{
\item{dir}{The subdirectory within Dropbox.}
}
\value{
A string containing the full path.
}
\description{
Returns the path to a specified folder within the Dropbox directory.
}
