% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years-life-lost.R
\name{expand_icd}
\alias{expand_icd}
\title{Expand ICD Sequence into Individual Codes}
\usage{
expand_icd(x)
}
\arguments{
\item{x}{A character string representing an ICD sequence.}
}
\value{
A character string with the expanded ICD codes.
}
\description{
Expands a given ICD sequence (e.g., "A01:A05") into individual ICD codes.
}
