% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years-life-lost.R
\name{explode_icd}
\alias{explode_icd}
\title{Explode ICD Group into Individual Codes}
\usage{
explode_icd(x)
}
\arguments{
\item{x}{A character string representing an ICD group.}
}
\value{
A character string with the expanded ICD codes.
}
\description{
Expands a given ICD group into individual ICD codes by recursively calling `expand_icd`.
}
