% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_pubmed}
\alias{extract_pubmed}
\title{Extract Data from PubMed File}
\usage{
extract_pubmed(file, what = c("TI", "AUTH", "SRC", "YEAR", "AB"))
}
\arguments{
\item{file}{Path to the PubMed file.}

\item{what}{Fields to extract (e.g., "TI", "AUTH", "SRC", "YEAR", "AB").}
}
\value{
A matrix containing the extracted fields.
}
\description{
Extracts specific fields from a PubMed file.
}
