% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{github}
\alias{github}
\title{Path to GitHub Folder}
\usage{
github(dir)
}
\arguments{
\item{dir}{The subdirectory within GitHub.}
}
\value{
A string containing the full path.
}
\description{
Returns the path to a specified folder within the GitHub directory.
}
