% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years-life-lost.R
\name{map_gbd}
\alias{map_gbd}
\title{Map GBD Level 4 Cause to Other Levels}
\usage{
map_gbd(x)
}
\arguments{
\item{x}{A vector of Level 4 GBD causes or ICD codes.}
}
\value{
A matrix with mappings to Level 4, Level 3, Level 2, and Level 1 categories.
}
\description{
Maps a Level 4 GBD cause to its corresponding Level 3, Level 2, and Level 1 categories.
}
