% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiplot}
\alias{multiplot}
\title{Plot Multiple ggplot2 Objects}
\usage{
multiplot(..., cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot2 objects to arrange.}

\item{cols}{Number of columns for the layout.}

\item{layout}{A matrix specifying the layout.}
}
\value{
None.
}
\description{
Arranges multiple ggplot2 objects on a single page.
}
