% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_best_inla_model.R
\name{predict_with_best_inla_model}
\alias{predict_with_best_inla_model}
\title{Predict new outcomes using the best INLA model}
\usage{
predict_with_best_inla_model(
  best_model_output,
  new_data,
  outcome_var,
  Ntrials = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{best_model_output}{A list containing the output from the `select_best_inla_model()` function. 
This list should include the best INLA model (`best_inla_model`), the model formula (`best_formula`),
and the original data used for model fitting (`original_data`).}

\item{new_data}{A data frame containing the new data for which predictions are needed.
This should match the structure of the original data, excluding the outcome variable.}

\item{outcome_var}{A character string specifying the name of the outcome variable in the original data.
The outcome variable in `new_data` will be set to `NA` for the purpose of prediction.}

\item{Ntrials}{Optional. Specifies the number of trials for binomial models. If `Ntrials = "N"`,
the function will use the "N" column from the new data. If not provided and required, an error will be thrown.}

\item{verbose}{Logical. If TRUE, provides verbose output from the `inla()` function during the model refitting.}

\item{...}{Additional arguments passed to the `inla()` function.}
}
\value{
A data frame containing the new data combined with predicted values and credible intervals (0.025quant, 0.5quant, 0.975quant).
        These values represent the fitted predictions along with their 95% credible intervals.
        The predictions are for the rows in `new_data`, and the original data rows are excluded from the returned output.
}
\description{
This function takes the output of the `select_best_inla_model()` function and a new dataset,
and performs predictions for the outcome variable using the best INLA model.
It refits the model with the original data combined with the new data, where the outcome variable
in the new data is set to NA.
}
\examples{
\dontrun{
# Assuming `best_model_output` is the result of `select_best_inla_model()`
# and `new_data` is a data frame for prediction:
predictions <- predict_with_best_inla_model(best_model_output, new_data, outcome_var = "outcome")
}

}
