% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_table}
\alias{prop_table}
\title{Proportional Table}
\usage{
prop_table(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional arguments passed to `table`.}
}
\value{
A proportional table.
}
\description{
Returns a proportional table for a given vector.
}
