% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet_source}
\alias{quiet_source}
\title{Source a File Quietly}
\usage{
quiet_source(file)
}
\arguments{
\item{file}{The file path of the R script.}
}
\value{
None.
}
\description{
Sources an R script without printing or plotting.
}
