% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readxl}
\alias{readxl}
\title{Read Excel File}
\usage{
readxl(...)
}
\arguments{
\item{...}{Arguments passed to `read_excel`.}
}
\value{
A data frame with cleaned column names.
}
\description{
Reads an Excel file and returns a data frame.
}
