% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years-life-lost.R
\name{rsle}
\alias{rsle}
\title{Residual Standard Life Expectancy (2019)}
\usage{
rsle(x)
}
\arguments{
\item{x}{Age(s) to calculate residual life expectancy for.}
}
\value{
Residual life expectancy for the given age(s).
}
\description{
Computes the residual life expectancy for a given age using the GBD 2019 life expectancy table.
}
