% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years-life-lost.R
\name{rsle2021}
\alias{rsle2021}
\title{Residual Standard Life Expectancy (2021)}
\usage{
rsle2021(x)
}
\arguments{
\item{x}{Age(s) to calculate residual life expectancy for.}
}
\value{
Residual life expectancy for the given age(s).
}
\description{
Computes the residual life expectancy for a given age using the GBD 2021 life expectancy table.
}
