% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_specials}
\alias{sanitize_specials}
\title{Translate Special Characters}
\usage{
sanitize_specials(char, type = c("html", "latex"))
}
\arguments{
\item{char}{A character string to sanitize.}

\item{type}{The target format ("html" or "latex").}
}
\value{
A sanitized character string.
}
\description{
Converts special characters into HTML or LaTeX representations.
}
