% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_inla_model.R
\name{select_best_inla_model}
\alias{select_best_inla_model}
\title{Select the best INLA model based on WAIC or DIC}
\usage{
select_best_inla_model(
  data,
  model_list,
  likelihood = "zeroinflatedbinomial0",
  metric = c("WAIC", "DIC"),
  Ntrials = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame or data.table containing the data for model fitting.}

\item{model_list}{A list of model formulas. Each model formula should be a valid formula that can be used in INLA.}

\item{likelihood}{A character string specifying the likelihood model to be used. Defaults to "zeroinflatedbinomial0".
Accepted values are for binomial and poisson families (e.g., "binomial", "poisson").}

\item{metric}{A character string specifying the metric for model comparison. Either "WAIC" or "DIC". Defaults to "WAIC".}

\item{Ntrials}{Optional. Specifies the number of trials for binomial models. If set to "N", it will use the "N" column from the data.}

\item{verbose}{Logical. If TRUE, provides verbose output during the INLA model fitting process.}

\item{...}{Additional arguments passed to the `inla()` function.}
}
\value{
A list containing:
  \item{model_comparisons}{A tibble with comparison of the models based on the chosen metric (WAIC or DIC).}
  \item{selected_model}{The name of the selected best-fitting model.}
  \item{best_inla_model}{The INLA model object for the best-fitting model.}
  \item{best_formula}{The formula used for the best-fitting model.}
  \item{original_data}{The original data used for fitting the models.}
}
\description{
This function fits multiple INLA models provided in a list and compares them based on WAIC or DIC.
It selects the best-fitting model and returns both the comparison of models and the best model itself.
It can handle binomial and poisson likelihoods, and optionally takes Ntrials for grouped binomial data.
}
\examples{
\dontrun{
# Define a list of models
model_list <- list(
  FIXED = ID ~ 1 + YEAR + SEX + AGEGR + REGION,
  TWOxWAYxYEAR = ID ~ 1 + YEAR*AGEGR + YEAR*REGION + YEAR*SEX
)

# Run the model selection
best_model_output <- select_best_inla_model(data, model_list, likelihood = "binomial", metric = "WAIC")

# Access the best model
best_inla_model <- best_model_output$best_inla_model
}

}
